refactor(l5): move observability context, with-logging and the logging barrel from src/ into server/observability (issue 864)
Three modules leave src/ for the server/ layer they already belonged to:

  src/observability/context.ts      -> server/observability/context.ts
  src/observability/with-logging.ts -> server/observability/with-logging.ts
  src/observability/index.ts        -> server/observability/log-surface.ts

The barrel does not keep its basename. server/observability/index.ts already
exists and is the ServerModuleBoundary descriptor for the module -- it declares
OBSERVABILITY_BOUNDARY and is re-exported by server/index.ts. Landing the
logging barrel on that same basename would have collided with the descriptor,
so the barrel arrives as log-surface.ts: the logging surface domain code
imports, distinct from the boundary declaration that describes the module.

Relative specifiers inside the moved files follow rule M3. The logger itself
has not moved, so the re-export and both imports now name ../../src/logger.ts.
context.ts and with-logging.ts sit beside the barrel, so ./context.ts and
./with-logging.ts are unchanged.

Each old path stays as a two-line rule-M2 shim re-exporting the new location,
which is what leaves every src/ importer's line untouched. server/domain/
sharing.ts is a server/ importer, so it names the real module directly at
../observability/log-surface.ts rather than routing back through src/.

OBSERVABILITY_BOUNDARY's owns/excludes lists hold prose responsibility phrases
("logger configuration", "correlation context"), not file basenames, so the
three new basenames were not added; nothing in that descriptor is keyed to a
filename.

src/observability/observability.test.ts stays in src/ untouched under the M1
exception and exercises the moved modules through the shims.

Verified: bunx tsc --noEmit exit 0; oxlint --deny-warnings on all seven staged
paths exit 0; bun run test:isolated over observability.test.ts,
langfuse-tracing.test.ts, trace-logger-injection.test.ts and sharing.test.ts
175 pass / 0 fail; scripts/done-means/864-moved-out-of-src.sh judged=28 shims
PASS; src-runtime-closure 64 -> 62 entries.
